test: normalize pagination results beyond search
